Move max_dop and spill_mem to proc RuntimeInfo
Updated proc.  Spill_mem default to auto configure from
system mem and max_dop.

+++ cmd/mo-service/main.go
@@ -78,7 +78,7 @@ func init() {
 
 func main() {
 	if *maxProcessor > 0 {
-		system.SetGoMaxProcs(*maxProcessor)
+		system.SetGoMaxProcs(int32(*maxProcessor))
 	}
 
 	flag.Parse()
